Send key stroke to the game on hotkey pressed
I thought that it might be good to be able to interact with markers (nodes, lore, etc...) and still be able to interact with the game.  It would then be able to harvest a node and update its timer on the website with a single key stroke.
I did not tested this since I don't have an Overwolf developer account but if I understood the documentation correctly, it should work.